added list class to simpleList program
