+++ simpleList/Makefile
@@ -10,8 +10,8 @@ LDFLAGS =
 # Makefile settings - Can be customized.
 APPNAME = simpleList
 EXT = .cpp
-SRCDIR = /Users/monish/personal-workspace/udacity-cpp-nanodegree/simpleList/
-OBJDIR = /Users/monish/personal-workspace/udacity-cpp-nanodegree/simpleList/
+SRCDIR = src
+OBJDIR = obj
 
 ############## Do not change anything from here downwards! #############
 SRC = $(wildcard $(SRCDIR)/*$(EXT))
