+++ simpleList/list.d
@@ -1,0 +1,1 @@
+obj/list.o: src/list.cpp src/include/list.h
