+++ simpleList/src/include/list.h
@@ -1,0 +1,28 @@
+#include <iostream>
+#include <vector>
+#include <string>
+
+using namespace std;
+
+class List {
+    private:
+    protected:
+
+    public:
+
+    List(){
+        // Constructor
+    }
+
+    ~List(){
+        // Destructor
+    }
+    vector<string> list;
+    string name;
+
+    void print_menu();
+    void print_list();
+    void add_item();
+    void delete_item();    
+ 
+};
