+++ simpleList/src/list.cpp
@@ -1,31 +1,6 @@
-#include <iostream>
-#include <vector>
-#include <string>
+#include "include/list.h"
 
-using namespace std;
-
-void print_menu(string name);
-void print_list();
-void add_item();
-void delete_item();
-
-vector<string> list;
-string name;
-
-int main(int arg_count, char *args[]){
-
-    if(arg_count>1){
-        name = string(args[1]);
-        print_menu(name);
-    }
-    else{
-        cout << "Username not supplied...exiting the program" << endl;
-    }
-
-    return 0;
-}
-
-void print_menu(string name){
+void List::print_menu(){
     int choice;
 
     cout << "***************************\n";
@@ -54,7 +29,7 @@ void print_menu(string name){
     }
 }
 
-void print_list(){
+void List::print_list(){
     cout<< "\n\n\n\n";
     cout<< "**** Printing List ****\n";
 
@@ -67,7 +42,7 @@ void print_list(){
     cin>> choice;
 
     if(choice == 'M' || choice == 'm'){
-        print_menu(name);
+        print_menu();
     }
     else{
         cout<< "Invalid choice. Quitting...\n\n";
@@ -75,7 +50,7 @@ void print_list(){
 }
 
 
-void add_item(){
+void List::add_item(){
     cout<< "\n\n\n\n";
     cout<< "**** Add Item ****\n";
     cout<< "Type an item and press enter: ";
@@ -88,10 +63,10 @@ void add_item(){
     cout<<"Successfully added item to the list \n\n\n";
     cin.clear();
 
-    print_menu(name);
+    print_menu();
 }
 
-void delete_item(){
+void List::delete_item(){
     cout<< "**** Delete Item ****\n";
     cout<< "Select an item index number to delete: \n";
 
@@ -104,8 +79,8 @@ void delete_item(){
         cout<< "No items to delete.\n";
     }
 
-    int index;
+    int index;   
     cin>> index;
 
-    print_menu(name);
+    print_menu();
 }
